src/main/java/com/example/demo/Service/MilestoneService.java
package com.example.demo.Service;

import com.example.demo.Models.Milestone;
import com.example.demo.Repository.MilestoneRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class MilestoneService {

    @Autowired
    private MilestoneRepository milestoneRepository;

    public Milestone createMilestone(Milestone milestone) {
        return milestoneRepository.save(milestone);
    }

    public List<Milestone> getAllMilestones() {
        return milestoneRepository.findAll();
    }

    public Milestone getMilestoneById(int milestoneId) {
        return milestoneRepository.findByMilestoneId(milestoneId);
    }
}

